Add support for UCB hardfloat (tests passed)

+++ src/test/scala/chiseltorch/TensorTest.scala
@@ -3,7 +3,9 @@ package tensor
 import chisel3._
 import chisel3.iotesters._
 import org.scalatest.{FlatSpec, Matchers}
-import tensor.dtypes.{DType, NDSeq}
+import chiseltorch.dtypes.DType
+import chiseltorch._
+import chiseltorch.common.NDSeq
 
 class TensorIntAssignmentTestModule extends Module {
     val tensor_temp = Tensor.empty(Seq(2, 2, 2), () => new dtypes.UInt(8.W))
@@ -38,15 +40,15 @@ class TensorIntAssignmentPeekPokeTest(c: TensorIntAssignmentTestModule) extends 
 
 
 class TensorFloatTestModule extends Module {
-    val tensor_temp = Tensor.empty(Seq(2, 2), () => new dtypes.Float(9, 7))
+    val tensor_temp = Tensor.empty(Seq(2, 2), () => new dtypes.Float(8, 9))
     val io = IO(new Bundle {
         val in1 = Input(tensor_temp.asVecType)
         val in2 = Input(tensor_temp.asVecType)
         val out = Output(tensor_temp.asVecType)
     })
 
-    val t1 = Tensor.Wire(Tensor.empty(Seq(2, 2), () => new dtypes.Float(9, 7)))
-    val t2 = Tensor.Wire(Tensor.empty(Seq(2, 2), () => new dtypes.Float(9, 7)))
+    val t1 = Tensor.Wire(Tensor.empty(Seq(2, 2), () => new dtypes.Float(8, 9)))
+    val t2 = Tensor.Wire(Tensor.empty(Seq(2, 2), () => new dtypes.Float(8, 9)))
     t1 := io.in1
     t2 := io.in2
     val t3 = t1 + t2
@@ -57,7 +59,6 @@ class TensorFloatTestModule extends Module {
 
 class TensorFloatPeekPokeTest(c: TensorFloatTestModule) extends PeekPokeTester(c) {
     println("Hello World")
-
 }
